// _frontend/models/cloud-mining-subscription-model.ts
export default interface CloudMiningSubscriptionModel {
  m3_price?: string | null;
  m3_daily_range?: string | null;
  m3_total_range?: string | null;
  m3_overview?: string | null;

  m6_price?: string | null;
  m6_daily_range?: string | null;
  m6_total_range?: string | null;
  m6_overview?: string | null;

  m12_price?: string | null;
  m12_daily_range?: string | null;
  m12_total_range?: string | null;
  m12_overview?: string | null;
}